:sparkles:  feat: refactor shuffle
