+++ MIGRATION.md
@@ -190,3 +190,8 @@ No breaking changes.
       }
     }
     ```
+
+
+### @algorithm.ts/knuth-shuffle
+
+1. Renamed to `@algorithm.ts/shuffle`.
