+++ packages/shuffle/src/index.ts
@@ -1,0 +1,6 @@
+import { knuthShuffle } from './knuth'
+
+export * from './knuth'
+export * from './util'
+
+export default knuthShuffle
